Don't use deprecated type

+++ tests/pyccel/scripts/functions.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 import numpy as np
 
-def incr_(x : np.int):
+def incr_(x : np.int32):
     x = x + 2
     return x
